# Conflicts:
#	src/main/java/com/s63b/domain/Result.java

+++ src/main/java/com/s63b/domain/Pol.java
@@ -2,9 +2,6 @@ package com.s63b.domain;
 
 import javax.persistence.*;
 
-/**
- * Created by bramd on 7-3-2017.
- */
 @Entity
 @NamedQueries({
         @NamedQuery(name = "Pol.getPolls", query = "SELECT pol FROM Pol AS pol WHERE licensePlate = :licensePlate"),
@@ -16,13 +13,13 @@ public class Pol {
     private String id;
 
     private String licensePlate;
-    private float lat;
-    private float lng;
+    private double lat;
+    private double lng;
     private long timestampMillis;
 
     public Pol() {}
 
-    public Pol(String licensePlate, float lat, float lng, long timestampMillis) {
+    public Pol(String licensePlate, double lat, double lng, long timestampMillis) {
         this.licensePlate = licensePlate;
         this.lat = lat;
         this.lng = lng;
@@ -37,11 +34,11 @@ public class Pol {
         return licensePlate;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public float getLng() {
+    public double getLng() {
         return lng;
     }
